+++ lib/flutter_view_controller.dart
@@ -5,3 +5,4 @@ export 'package:plug/plug.dart';
 export 'package:sizer_plus/sizer_plus.dart';
 export 'src/view.dart';
 export 'src/navigator_monitor.dart';
+export 'src/ui_component.dart';