+++ CHANGELOG.md
@@ -178,3 +178,4 @@
 * REINTEGRATED Plug FOR DEVELOPMENT FACILITIES
 * CREATED Anima class for integrated animations
 * CREATED Sizer.heightResponsible, Sizer.widthResponsible, Sizer.fraction(heights, widths) for breakpoints percentages widths and heights.
+* INTERNALIZED Plug _isConnected BEHAVIOR